+++ io7m-jcamera-core/src/main/java/com/io7m/jcamera/JCameraSphericalAngularIntegrator.java
@@ -47,8 +47,8 @@ public final class JCameraSphericalAngularIntegrator implements
     this.camera = NullCheck.notNull(in_camera, "Camera");
     this.input = NullCheck.notNull(in_input, "Input");
 
-    this.maximum_speed_heading = (float) (2 * Math.PI);
-    this.maximum_speed_incline = (float) (2 * Math.PI);
+    this.maximum_speed_heading = (float) (2.0 * Math.PI);
+    this.maximum_speed_incline = (float) (2.0 * Math.PI);
     this.acceleration_heading = this.maximum_speed_heading / 2.0f;
     this.acceleration_incline = this.maximum_speed_incline / 2.0f;
     this.drag_incline = 0.05f;
@@ -62,7 +62,7 @@ public final class JCameraSphericalAngularIntegrator implements
     final float drag,
     final float time)
   {
-    return (float) (f * Math.pow(drag, time));
+    return (float) ((double) f * Math.pow((double) drag, (double) time));
   }
 
   /**
@@ -154,7 +154,7 @@ public final class JCameraSphericalAngularIntegrator implements
   {
     this.acceleration_heading =
       (float) RangeCheck.checkGreaterDouble(
-        a,
+        (double) a,
         "Acceleration",
         0.0,
         "Minimum acceleration");
@@ -167,7 +167,7 @@ public final class JCameraSphericalAngularIntegrator implements
     this.drag_heading =
       (float) RangeCheck.checkGreaterEqualDouble(
         RangeCheck
-          .checkLessEqualDouble(d, "Drag factor", 1.0, "Maximum drag"),
+          .checkLessEqualDouble((double) d, "Drag factor", 1.0, "Maximum drag"),
         "Drag factor",
         0.0,
         "Minimum drag");
@@ -179,7 +179,7 @@ public final class JCameraSphericalAngularIntegrator implements
   {
     this.maximum_speed_heading =
       (float) RangeCheck.checkGreaterEqualDouble(
-        s,
+        (double) s,
         "Speed limit",
         0.0,
         "Minimum limit");
@@ -191,7 +191,7 @@ public final class JCameraSphericalAngularIntegrator implements
   {
     this.acceleration_incline =
       (float) RangeCheck.checkGreaterDouble(
-        a,
+        (double) a,
         "Acceleration",
         0.0,
         "Minimum acceleration");
@@ -204,7 +204,7 @@ public final class JCameraSphericalAngularIntegrator implements
     this.drag_incline =
       (float) RangeCheck.checkGreaterEqualDouble(
         RangeCheck
-          .checkLessEqualDouble(d, "Drag factor", 1.0, "Maximum drag"),
+          .checkLessEqualDouble((double) d, "Drag factor", 1.0, "Maximum drag"),
         "Drag factor",
         0.0,
         "Minimum drag");
@@ -216,7 +216,7 @@ public final class JCameraSphericalAngularIntegrator implements
   {
     this.maximum_speed_incline =
       (float) RangeCheck.checkGreaterEqualDouble(
-        s,
+        (double) s,
         "Speed limit",
         0.0,
         "Minimum limit");
