+++ io7m-jcamera-core/src/main/java/com/io7m/jcamera/JCameraSphericalLinearIntegrator.java
@@ -40,6 +40,7 @@ public final class JCameraSphericalLinearIntegrator implements
   private       float                 zoom_acceleration;
   private       float                 zoom_drag;
   private       float                 zoom_maximum_speed;
+
   private JCameraSphericalLinearIntegrator(
     final JCameraSphericalType in_camera,
     final JCameraSphericalInput in_input)
@@ -66,7 +67,7 @@ public final class JCameraSphericalLinearIntegrator implements
     final float drag,
     final float time)
   {
-    return (float) (f * Math.pow(drag, time));
+    return (float) ((double) f * Math.pow((double) drag, (double) time));
   }
 
   /**
@@ -206,7 +207,7 @@ public final class JCameraSphericalLinearIntegrator implements
   {
     this.target_acceleration =
       (float) RangeCheck.checkGreaterDouble(
-        a,
+        (double) a,
         "Acceleration",
         0.0,
         "Minimum acceleration");
@@ -219,7 +220,7 @@ public final class JCameraSphericalLinearIntegrator implements
     this.target_drag =
       (float) RangeCheck.checkGreaterEqualDouble(
         RangeCheck
-          .checkLessEqualDouble(f, "Drag factor", 1.0, "Maximum drag"),
+          .checkLessEqualDouble((double) f, "Drag factor", 1.0, "Maximum drag"),
         "Drag factor",
         0.0,
         "Minimum drag");
@@ -231,7 +232,7 @@ public final class JCameraSphericalLinearIntegrator implements
   {
     this.target_maximum_speed =
       (float) RangeCheck.checkGreaterEqualDouble(
-        s,
+        (double) s,
         "Speed limit",
         0.0,
         "Minimum limit");
@@ -243,7 +244,7 @@ public final class JCameraSphericalLinearIntegrator implements
   {
     this.zoom_acceleration =
       (float) RangeCheck.checkGreaterDouble(
-        a,
+        (double) a,
         "Acceleration",
         0.0,
         "Minimum acceleration");
@@ -256,7 +257,7 @@ public final class JCameraSphericalLinearIntegrator implements
     this.zoom_drag =
       (float) RangeCheck.checkGreaterEqualDouble(
         RangeCheck
-          .checkLessEqualDouble(f, "Drag factor", 1.0, "Maximum drag"),
+          .checkLessEqualDouble((double) f, "Drag factor", 1.0, "Maximum drag"),
         "Drag factor",
         0.0,
         "Minimum drag");
@@ -268,7 +269,7 @@ public final class JCameraSphericalLinearIntegrator implements
   {
     this.zoom_maximum_speed =
       (float) RangeCheck.checkGreaterEqualDouble(
-        s,
+        (double) s,
         "Speed limit",
         0.0,
         "Minimum limit");
