+++ io7m-jcamera-core/src/main/java/com/io7m/jcamera/JCameraSphericalMouseRegion.java
@@ -42,17 +42,17 @@ public final class JCameraSphericalMouseRegion
 
     this.height =
       (float) RangeCheck.checkGreaterEqualDouble(
-        in_height,
+        (double) in_height,
         "Height",
-        2.0f,
+        2.0,
         "Minimum height");
     this.center_y = this.height / 2.0f;
 
     this.width =
       (float) RangeCheck.checkGreaterEqualDouble(
-        in_width,
+        (double) in_width,
         "Width",
-        2.0f,
+        2.0,
         "Minimum width");
     this.center_x = this.width / 2.0f;
   }
@@ -94,8 +94,8 @@ public final class JCameraSphericalMouseRegion
   }
 
   /**
-   * Get the normalized coordinates for the given screen coordinates <code>(x,
-   * y)</code>.
+   * Get the normalized coordinates for the given screen coordinates {@code (x,
+   * y)}.
    *
    * @param x   The x coordinate
    * @param y   The y coordinate
