+++ io7m-jcamera-examples-jogl/src/main/java/com/io7m/jcamera/examples/jogl/ExampleFPSStyleGLListener.java
@@ -110,15 +110,15 @@ public final class ExampleFPSStyleGLListener implements GLEventListener
 
     final long time_now = System.nanoTime();
     final long time_diff = time_now - this.time_then;
-    final double time_diff_s = time_diff / 1000000000.0;
+    final double time_diff_s = (double) time_diff / 1000000000.0;
     this.time_accum = this.time_accum + time_diff_s;
     this.time_then = time_now;
 
     final float sim_delta = this.sim.getDeltaTime();
-    while (this.time_accum >= sim_delta) {
+    while (this.time_accum >= (double) sim_delta) {
       this.snap_prev = this.snap_curr;
       this.snap_curr = this.sim.integrate();
-      this.time_accum -= sim_delta;
+      this.time_accum -= (double) sim_delta;
     }
 
     /**
@@ -127,7 +127,7 @@ public final class ExampleFPSStyleGLListener implements GLEventListener
      * states.
      */
 
-    final float alpha = (float) (this.time_accum / sim_delta);
+    final float alpha = (float) (this.time_accum / (double) sim_delta);
     final JCameraFPSStyleSnapshot snap_interpolated =
       JCameraFPSStyleSnapshot.interpolate(
         this.snap_prev,
@@ -156,8 +156,8 @@ public final class ExampleFPSStyleGLListener implements GLEventListener
   {
     this.mouse_region.set(JCameraFPSStyleMouseRegion.newRegion(
       JCameraScreenOrigin.SCREEN_ORIGIN_TOP_LEFT,
-      width,
-      height));
+      (float) width,
+      (float) height));
     this.renderer.reshape(width, height);
   }
 }
